backend/requirements.txt
cffi==1.13.2
cryptography==2.8
Django==2.2.8
django-cors-headers==3.1.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
pycparser==2.19
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
